+++ cordformation/src/test/kotlin/net/corda/plugins/BaseformTest.kt
@@ -2,10 +2,6 @@ package net.corda.plugins
 
 import com.typesafe.config.Config
 import com.typesafe.config.ConfigFactory
-import net.corda.core.serialization.SerializationContext
-import net.corda.serialization.internal.CordaSerializationMagic
-import net.corda.serialization.internal.amqp.AbstractAMQPSerializationScheme
-import net.corda.serialization.internal.amqp.amqpMagic
 import org.assertj.core.api.Assertions.assertThat
 import org.gradle.testkit.runner.GradleRunner
 import org.junit.jupiter.api.io.TempDir
@@ -18,10 +14,15 @@ open class BaseformTest {
     @TempDir
     lateinit var testProjectDir: Path
 
+    private lateinit var testRoot: String
+
     companion object {
-        const val cordaFinanceWorkflowsJarName = "corda-finance-workflows-4.3"
-        const val cordaFinanceContractsJarName = "corda-finance-contracts-4.3"
-        const val localCordappJarName = "locally-built-cordapp"
+        private const val cordaVersion = "1618234704308-rc"
+        const val cordaBundleVersion = "1.0.$cordaVersion"
+        const val cordaReleaseVersion = "5.0.$cordaVersion"
+        const val cordaFinanceWorkflowsCpkName = "corda-finance-workflows-$cordaBundleVersion-cordapp"
+        const val cordaFinanceContractsCpkName = "corda-finance-contracts-$cordaBundleVersion-cordapp"
+        const val localCordappCpkName = "locally-built-cordapp"
         const val bankNodeName = "BankOfCorda"
         const val notaryNodeName = "NotaryService"
         const val notaryNodeUnitName = "OrgUnit"
@@ -31,7 +32,7 @@ open class BaseformTest {
 
     fun getStandardGradleRunnerFor(
             buildFileResourceName: String,
-            taskName: String = "deployNodes",
+            taskName: String,
             vararg extraArgs: String
     ): GradleRunner {
         val isKotlin = buildFileResourceName.endsWith(".kts")
@@ -44,7 +45,10 @@ open class BaseformTest {
         val buildFile = testProjectDir.resolve(buildScriptName)
 
         createBuildFile(buildFileResourceName, buildFile)
-        installResource("settings.gradle")
+        testRoot = buildFileResourceName.replaceAfterLast('/', "")
+        if (!installResource("$testRoot/settings.gradle")) {
+            installResource("settings.gradle")
+        }
         installResource("repositories.gradle")
         installResource("gradle.properties")
         installResource("postgres.gradle")
@@ -55,22 +59,34 @@ open class BaseformTest {
                 .withPluginClasspath()
     }
 
-    private fun createBuildFile(buildFileResourceName: String, buildFile: Path): Long {
-        return javaClass.getResourceAsStream(buildFileResourceName)?.use { s ->
-            Files.copy(s, buildFile)
-        } ?: throw NoSuchFileException(buildFileResourceName)
+    private fun createBuildFile(buildFileResourceName: String, buildFile: Path) {
+        if (copyResourceTo(buildFile, buildFileResourceName) < 0) {
+            throw NoSuchFileException(buildFileResourceName)
+        }
     }
 
-    fun installResource(resourceName: String) {
-        val buildFile = testProjectDir.resolve(resourceName.substring(1 + resourceName.lastIndexOf('/')))
-        javaClass.classLoader.getResourceAsStream(resourceName)?.use { input ->
-            Files.copy(input, buildFile)
-        } ?: throw NoSuchFileException(resourceName)
+    fun installResource(resourceName: String): Boolean {
+        val filePath = resourceName.removePrefix(testRoot).dropWhile { it == '/' }
+        val fileIdx = filePath.lastIndexOf('/')
+        val directory = if (fileIdx == -1) {
+            testProjectDir
+        } else {
+            Files.createDirectories(testProjectDir.resolve(filePath.substring(0, fileIdx)))
+        }
+        val buildFile = directory.resolve(filePath.substring(fileIdx + 1))
+        return copyResourceTo(buildFile, resourceName) >= 0
+    }
+
+    private fun copyResourceTo(target: Path, resourceName: String): Long {
+        return javaClass.classLoader.getResourceAsStream(resourceName)?.use { input ->
+            Files.copy(input, target)
+        } ?: -1
     }
 
     fun getNodeLogFile( nodeName: String, fileName: String ): Path = Paths.get(testProjectDir.toAbsolutePath().toString(), "build", "nodes", nodeName, "logs", fileName)
-    fun getNodeCordappJar(nodeName: String, cordappJarName: String): Path = Paths.get(testProjectDir.toAbsolutePath().toString(), "build", "nodes", nodeName, "cordapps", "$cordappJarName.jar")
-    fun getNodeCordappConfig(nodeName: String, cordappJarName: String): Path = Paths.get(testProjectDir.toAbsolutePath().toString(), "build", "nodes", nodeName, "cordapps", "config", "$cordappJarName.conf")
+    fun getNodeCordappDir(nodeName: String): Path = Paths.get(testProjectDir.toAbsolutePath().toString(), "build", "nodes", nodeName, "cordapps")
+    fun getNodeCordappCpk(nodeName: String, cordappCpkName: String): Path = getNodeCordappDir(nodeName).resolve("$cordappCpkName.cpk")
+    fun getNodeCordappConfig(nodeName: String, cordappCpkName: String): Path = getNodeCordappDir(nodeName).resolve("config").resolve("$cordappCpkName.conf")
     fun getNetworkParameterOverrides(nodeName: String): Path = Paths.get(testProjectDir.toAbsolutePath().toString(), "build", "nodes", nodeName, "network-parameters")
     fun getNodeConfigFile(nodeName: String): Path = Paths.get(testProjectDir.toAbsolutePath().toString(), "build", "nodes", nodeName, "node.conf")
 
@@ -80,12 +96,12 @@ open class BaseformTest {
         return ConfigFactory.parseFile(configFile.toFile())
     }
 
-    class AMQPParametersSerializationScheme : AbstractAMQPSerializationScheme(emptyList()) {
-        override fun rpcClientSerializerFactory(context: SerializationContext) = throw UnsupportedOperationException()
-        override fun rpcServerSerializerFactory(context: SerializationContext) = throw UnsupportedOperationException()
-
-        override fun canDeserializeVersion(magic: CordaSerializationMagic, target: SerializationContext.UseCase): Boolean {
-            return magic == amqpMagic && target == SerializationContext.UseCase.P2P
-        }
-    }
+//    class AMQPParametersSerializationScheme : AbstractAMQPSerializationScheme() {
+//        override fun rpcClientSerializerFactory(context: SerializationContext) = throw UnsupportedOperationException()
+//        override fun rpcServerSerializerFactory(context: SerializationContext) = throw UnsupportedOperationException()
+//
+//        override fun canDeserializeVersion(magic: CordaSerializationMagic, target: SerializationContext.UseCase): Boolean {
+//            return magic == amqpMagic && target == SerializationContext.UseCase.P2P
+//        }
+//    }
 }