+++ changelog.md
@@ -10,6 +10,7 @@
 * `cordformation`: Replace `Closure` with `Action` in the DSL.
 * `cordformation`: Annotate `Baseform.signing` as `@Nested` instead of `@Input`.
 * `cordapp-cpk`: Generate CPK format CorDapps. Requires Gradle 6.6+.
+* `cordformation`: Update to deploy CPKs for Corda 5.
 
 ## Version 5